Add Endogenous Retirements to Temoa
This commit adds an endogenous retirement feature to Temoa.

Users can add a list of technologies for which they would like to enable
endogenous retirements to the 'tech_retirement' set.

Prior to this commit, the decision variables of Temoa were:
- Capacity additions
- Operational decision at the time step level

This commit adds an additional decision variable of
- Retirement decisions

Keep in mind that enabling the endogenous retirement feature will add to model
size, complexity and compute time. As a result, users are advised to only include
technologies for which they explicitly want to enable endogenous retirements to the
'tech_retirement' set.